requirements.txt
asgiref==3.8.1
bpy==4.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
Cython==3.0.10
dataclasses-json==0.6.6
Deprecated==1.2.14
dill==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
fake-bpy-module==20240514
idna==3.7
Js2Py==0.74
kad_py_test==0.0.98
marshmallow==3.21.2
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pygltflib==1.16.2
pyjsparser==2.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
wrapt==1.16.0
zstandard==0.22.0
